chore(Storybook): Gather the index page example content in one module
The News Overview, Events Overview and Search Results pages each kept a
data module of their own, and two of them held a list of districts that
exampleContent already had. Move the types and the fixtures to
exampleContent and take the districts from the list that was there.

The News Overview list was the odd one out: unsorted, and naming Weesp
rather than Stadsgebied Weesp. The shared list wins on both counts, so its
filter now reads alphabetically and uses the official name.
